backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai>=0.8.0
firebase-admin
google-cloud-firestore
google-cloud-storage
google-cloud-bigquery
scipy
numpy
pandas
httpx
anthropic
reportlab
python-dotenv
pydantic>=2.0
